Improve scaling and introduce flexible scaling mode (distorts text).

+++ src/main/java/gigaherz/guidebook/guidebook/IBookGraphics.java
@@ -11,7 +11,7 @@ public interface IBookGraphics
 {
     boolean refreshScalingFactor();
 
-    float getScalingFactor();
+    double getScalingFactor();
 
     boolean canGoBack();
